Notify Preview render before applying zoom-to-fit
Fixes a race condition between DesignSurface.afterResize
and Preview.onAfterRender.
The two callbacks called restoreZoomOrZoomToFit causing a race condition
and a not correct calculation of zoom-to-fit.

This fix adds a synchronized notification of the two objects that will
call restoreZoomOrZoomToFit when both Preview and DesignSurface provides
respectively the both content size and window size.

Fixes: 369957614
Test: DesignSurfaceTest
Change-Id: Icaf643ea564a60faf0899249582ceb50ab72efb8

GitOrigin-RevId: 3bf5b13325da9c2dbf96d6a97f14747c29da5384