Add sizeForScale to PositionableContent
sizeForScale will calculate total size for PositionableContent including
margins and content for target scale

sizeForScale = contentSize * scale + margin(scale)

* contentSize is the size of the content without any scaling.
contentSize is a "raw" size of the preview and is not changing unless
preview is updated. It should be scaled proportionally to get size to be
used in surface.

* margin is mostly fixed size, they are not scaled proportionally with
scale. margin covers the size of labels, toolbars or extra panels
surrounding the preview.

For SceneViewHeader sizeForScale is always fixed as it's not reacting
to scale changes.

Test: SceneViewHeaderTest
Bug: n/a
Change-Id: I17653b712f7dbcded2d79ebcee365ca9aead2364

GitOrigin-RevId: 67cd60a06b30ffbe958b54412948e34915f6ab32