Remove deprecated modification tracker
javaStructureModificationTracker is deprecated and will be removed. This
change replaces it with what it is returning internally anyway, so
there's no functional difference.

Also removing the TODO comment above. The associated bug is long closed,
and it's not a major performance issue for this cached value to be
invalidated on PSI updates. (It only applies to the calculated list of
fields on the BR class, which is not commonly used - the amount of
processing this will occur is minimal.)

Bug: n/a
Test: Covered by data binding tests
Change-Id: Ib5d76a7de9f9da7e0b3731a86b7e333d859df99d

GitOrigin-RevId: 7306ca7998c7181c79404b295d25d498d53f68e8