Adjust visibility and construction of maven index
MavenClassRegistry contains a bunch of top-level items that should be
private to the file (or in some cases, internal for test purposes but
functionally private).

Additionally, the class is effectively just a container but does a ton
of parsing on construction. Moving that logic to a `create` method to
more idiomatically show that computations are happening, which makes the
actual construction light-weight and allows us to potentially make the
create method suspending in the future.

Bug: 373342815
Test: Covered by existing
Change-Id: I26aac5be448709eed3c19f41f1dab4bbf57d0199

GitOrigin-RevId: 737f117e65b327ad74a1989277dbf68dba3df360