PSD: Apply specific treatment to kmp modules
PSD window does not support reading/updating KMP module data
so this is attempting to hide the data. Introducing a field to differentiate between regular android modules and kmp ones.
Note: this is a short term workaround and longer term we can introduce common abstract interface between the two and work based on that instead of a flag.

Bug: 371071431
Test: existing
Change-Id: Ic223ca267db11e55b2148073e41f3464ccdaf9ad

GitOrigin-RevId: ebb97a648ffaf509ed70971041d9d17e40531206