Handle older TileServiceViewAdapters
... when detecting animations.

In older androidx versions, this method does not exist, leading to an
exception being thrown. When this happens on the first render of the
preview the exception was being thrown before calling
previewViewModel.afterPreviewsRefreshed() in CommonPreviewRepresentation
leading to an indefinite "Initialization" screen instead of a preview.

This is because afterPreviewsRefreshed calls hasRendered.set(true)
under the hood which in turn is checked in
CommonPreviewViewModel#updateView in order to show the preview.

Fixes: 373681154
Test: added regression test
Change-Id: Iddc22ab48b9db95f631d854713b57076970d7c2b

GitOrigin-RevId: da0debd1efbe97f8c0ffef5a9193fcce4ebdf0c5