[DCL] Add ability to add/edit build properties based on software type
Fix cases when:
- we add new property and parent exist in build file
- we add new property and parent exist in setting file
- we add new property and direct parent does not exist
- we edit software type property and new property is created in build
file along with parent

There are three basic ideas here.
- The first is that we can track where the property physically came
from analyzing PsiElement.containingFile. It's not convenient
every time to go to parent, find child element = this and check
ItemElement status parameter - APPLIED or EXISTING.
- Applied is attribute of ElementItem that is how DSL element
includes it's child properties, when we parse build file and find
same block we need to substitute psiElement in existing DSL element
and make sure it's marked as non Applied (refreshAfterPsiSubstitution)
- As long as we work with deep tree - we need to mark all APPLIED
elements as applied because some of them may be rewritten in
build file.

Bug: 370972647
Test: AndroidModelTest
Change-Id: I81800f63ddf2fa32986984cf005dfd97f5b62603

GitOrigin-RevId: d10c1cc441206a166a508ef1622c9573b9653a68