Add GeminiPluginApi, a simpler API from android -> Gemini
The ml-api module exposes a fairly large set of APIs for use by the
android plugin in order to talk to the Gemini plugin. The same API
is used by both the functionality inside the android plugin and
within the Gemini plugin. This makes it very easy to develop functionality
in either of those plugins. However, this causes a number of problems,
esp. since we have to ship the aiplugin for IntelliJ as well.

This CL introduces an alternative, simpler API for use from the
android plugin. While simple, it does cover all the existing use cases
from the android plugin. The implementation of this API resides in
the aiplugin.android module (see companion CL), and it is mostly just
a thin wrapper on top of the existing implementation. This allows us
to, in a future CL, move the rest of the ml-api into the Gemini
plugin. This simplifies bundling and distribution, but it does mean
that if developers of the android plugin wish to use more complex
APIs, they'll have to do so within the aiplugin.android.

Follow up CLs will migrate all users of the existing StudioBot API
to the new GeminiPluginApi, and then move the StudioBot APIs out of
ml-api module.

Bug: 360363746
Test: added
Change-Id: I5b05c71c01a4db3a0ba2755d95ad84ea2cd44734

GitOrigin-RevId: 47ff948cf04399427467f6654148b235c1850db8