Use a record for storing rules instead of a proto.
This result in significant memory savings, because protos store an empty
proto list object for every empty repeated field. For a large project,
this resulted in a 28M duplicate such objects, wasting 2GB of RAM for a
very large project.

Using a record means that all empty fields share the same empty
ImmutableList object, and probably other optimizations too. For the same
large project, this cuts the size of the QuerySummary object by 50%, and
the whole heap size by about a quarter, a 3GB memory saving.

There is more work to do here. For example, there's no need to create or
store the `StoredRule` protos when parsing the query output, and
removing these is likely to give more memory savings. But this CL is
good to submit as is, especially since it gives good memory savings.

Test: Verified manually using heapdumps.
Bug: 374171123
Change-Id: I3569b5ee248d3aa3e7333c17b71c537a29a6811a

GitOrigin-RevId: 65057327eefe6e2b438a4b81a84f0572adeb8a55