[gradle][android] IDEA-347852 collect version catalogs in BuildModelContext without determining a base project directory
This commit sets up collecting version catalogs without usage of:
- `Project.guessProjectDir()`
- `Project.getBaseDir()` - deprecated
- `BuildModelContext.getBaseDirPath(project)` - uses not recommended `getBasePath()`
All these methods are not reliable and could bring unexpected results.

Before this commit, it was possible to encounter the bug when version catalogs were not collected correctly. The problem was in unexpected behavior of `guessProjectDir()` for a composite build with a linked project located outside the root project directory. I could not find a reliable alternative method. So, reworked the collecting version catalogs without determining a base project directory at all.

Probably, it makes sense to apply similar changes in `GradleVersionCatalogViewImpl` - take a value from the `from` property as a Virtual file, not as a String.

GitOrigin-RevId: c2ebabfa94ac3e92bedd785a6bfec59e251120e3