[gradle][android] IDEA-343916 fix collecting version catalogs defined settings.gradle (in Groovy) in composite builds
The problem: version catalogs with a custom location defined in settings.gradle were not added in myVersionCatalogFiles of BuildModelContext because `isFile()` have been returning `false` for a catalog like:
```
dependencyResolutionManagement {
    versionCatalogs {
        "customLibs" {
            from(files("customLocation/customLibs.versions.toml"))
...
```
As a consequence, resolving references in a build.gradle (in Groovy) to dependencies from `customLibs` was broken for included builds of composite builds.

The cause: parsing of settings files in Groovy and Kotlin works differently in GroovyDslParser and KotlinDslParser.

The solution: I was not sure which parser works more correctly, so adjusted `isFile()` to deal properly with the results of both.

P.S. These changes fix only resolving and navigation to version catalogs from Groovy build scripts of composite builds.
- For simple (non-composite) projects with Gradle scripts in Groovy, resolving worked even before these changes.
But in such cases, resolving relied on another logic implemented outside the android module (see VersionCatalogsModelBuilder).
- Navigation from Kotlin build scripts worked well in both composite and non-composite projects,
because the previous implementation of `isFile()` have been returning correct results for catalogs defined in settings files on Kotlin.

GitOrigin-RevId: 4111f8776c9c4f4622db69270600b490a3fc0e85