Fix zoom consistency of Gallery
Zoom wasn't consistent when switching from Grid
to Gallery.
The reason is that Gallery is not only a layout option but also a
mode.

When switching mode from/to Gallery we now reset the first render flag.
So that we can re-calculate DesignSurface sizes and restore the zoom
selected in the previous layout option.

Bug: 373348063
Test: existing tests should work
Change-Id: I7145ed46727f7e71aee07794bf9e0bf586e3a985

GitOrigin-RevId: d86bcafc27041f4be831ffcfa38b6af0b6b28de1