Don't throw from FutureCallback.onFailure
The context for running the callback depends too much on timing and
other environmental factors.  Instead, log a warning on failure and
otherwise silently do nothing: this is consistent with the platform
action that AndroidInferNullityAnnotationAction replaces, which also
restarts analysis after successfully adding dependencies and silently
does nothing on failure.

Fixes: 375355918
Test: existing, manual
Change-Id: Ic7c53c912504d8290f1ff4e0678a3abc7a7117ac

GitOrigin-RevId: 9d396bc6ff3080969e5fbcb5aa581022f744fd8d