[android][gradle] IDEA-361677 support version catalogs in included builds without a settings file.
Before this commit, version catalogs in an included build of a composite build were supported only if the included build had a settings file. If an included build didn't have its own settings file, BuildModel used a settings file from upper directories. This would be correct if the build was a subproject, but incorrect for an included build of a composite build.

There is a difference between a "subproject" and an "included build of a composite build":
1) In the root settings file:
  - `include("subproject")` registers subproject of a multi-project build.
  - `includeBuild("includedBuild")` registers an included build of a composite build.
2) A subproject uses version catalogs from the root settings file. An included build doesn't because it could have its own settings file and its own version catalogs.

GitOrigin-RevId: 0a9490348f6dac5330bf77a57708d9c9244d3504