Clear metadata cache when project is unloaded
The `RecipeUtils.getRecipeMetadata()` function converts a `Recipe` to a
`RecipeMetadata`. It caches the result in a hashmap, which is stored as
a static field of the `RecipeUtils` class. Since the `Project` object is
used as a cache key and cache entries are never evicted, the project is
leaked.

In this CL, we configure a Disposer to ensure the cache entries
associated with a project are cleared from the cache when the project is
closed.

The function has also been refactored to help avoid race condition
issues.

Bug: 288940571
Test: Added new unit test
Change-Id: I000c27808d59899b224dd46be0ba0f78d99f609b

GitOrigin-RevId: 37a91e18d0a02aaa594dd4da8d60bd39062c33b4