Re-arrange rule handling in BlazeQueryParser
to make it possible to extract the list of rule classes that are handled
by this class. This is needed to produce a query that filters out
unneeded rules as it the most straightforward query produces too large
outputs.

Note, in the future we might want to choose which rules to handle based
on other criteria and in this case we will also need to review how
the query selects rules to return to the IDE.

Bug: 362727747
Test: existing
Change-Id: Ia0bfbe0ac49e9b5e47669f77e1f90fe66e1e177e

GitOrigin-RevId: 306d712615fd8dc41621ec9dc1d5cbe965f24f3b