Add action to open Layout Inspector from Device Manager
This change adds an extension point to device manager that allows
clients to add actions to the overflow menu of a device.

Layout Inspector uses this extension point to add an action that
when selected enables standalone Layout Inspector, shows the tool
window and starts foreground process detection for the device.

The action is currently always disabled. It will be enabled
in a following change.

Test: NA
Bug: NA
Change-Id: I27e7edef7b2c481900e5d0b9742fa42cb94ba84f

GitOrigin-RevId: 2103ffe4cee1a98f5b5f721c24a326631cffad3c