Use file path as key for distinct
We read the entire contents of the file that a class is present in. It
might be possible to have multiple classes in the same file as well as
in the stacktrace. This will lead to adding the same file twice in the
prompt. Using file path adds the file only once.

Bug: N/A
Test: Updated
Change-Id: I2aa03bef3ba67e4dae9ed185e2560b2eaac70a8c

GitOrigin-RevId: f1006c3f7a5f6d68df7005049680a795ba56318d