Annotate RecipeUtils methods with target threads
The `RecipeUtils.getRecipeMetadata` function should be run on a worker
thread, as it can take a while to run for projects with lots of modules,
so let's annotate it with `@WorkerThread`. We'll do the work to ensure
it's always called from a worker thread in a separate CL with the same
topic.

The `RecipeUtils.execute` function must run in the EDT thread since it
performs a bunch of writes. It's already always called from this thread,
so need additional changes are needed.

Bug: 288940571
Test: Manual
Change-Id: I8073393fb75e21ac6eac40501bdab1594094cd07

GitOrigin-RevId: 269917e298bce2741264af2e071ac1ad58bb5563