Move IdeGoogleMavenRepository into its own module
IdeGoogleMavenRepository is the last remaining entity in android core's
com.android.tools.idea.gradle package, which is a thing which really
ought not to exist.  It cannot be moved directly to the Gradle project
system because of circular dependencies between it and the
"repositorySearch" subsystem; also, its overall functionality is not
Gradle-specific, pertaining as it does to published artifacts on the
web -- though its current implementation is, as many operations are
defined in terms of entities in Gradle's taxonomy, such as Dependency,
Component and Version.

Placing this into its own module is a luxury, but maybe a luxury we
can afford in order to write...

Fixes: 300434824
Test: existing
Change-Id: I818e440cc601734a8775631523f34f0feea25b21

GitOrigin-RevId: f1dd78323d173fa6369c849f2c253f2873ea32b6