Use python 3.8 supported type hints
The intellij_remote_platform repository_rule relies on calling python using the system python, which on studio-mac-arm is 3.8. I tried making the rule use a py_binary (so it uses the checked in python toolchain) but repository rules don't seem to support this. A better solution might be to generate the BUILD+spec files during check-in time, rather than having the repository_rule do it with python at build time.

Change-Id: Ib74ccf744a6767ad8fc9d281dab8f0b2e6e08ad6

GitOrigin-RevId: b705777d567d0ccd53b52225591cc99bace43ae7