Remove manual checks for cancellations in preview refresh
The current state of the refresh process is ready to properly handle
CancellationExceptions, and it already correctly handles those
caused by automatic cancellations triggered by the refresh manager.

However, user cancellations were being handled a bit different, as
the progress indicator is checked at different points during the
refresh process. This was acting as logic similar to the following:
"if a CancellationException is about to happen because the user
manually cancelled the refresh, then stop everything now and finish"

But this is not needed, and it causes some problems like sometimes
misinterpreting a user cancelled refresh as successful.

Bug: 335424569
Test: existing should pass and manually checked preview tools
Change-Id: Ia514a3dd965a089c88d49ade589202808809a2aa

GitOrigin-RevId: 27dd8867ec6889a36662a0f464893f96826fd2dd