Migrate 'Welcome' step to use new wizard
The UI has been copied from the deprecated `FirstRunWelcomeStep` class
and has been adapted to work with the `ModelWizardStep` class. It's a
basic UI that simply displays text and images, so was a straightforward
migration.

You can force the new wizard to launch when running Android Studio
locally by setting the `NPW_FIRST_RUN_WIZARD` and `NPW_FIRST_RUN_SHOW`
flags to `true` in `StudioFlags.java`.

Bug: 311256502
Test: WelcomeScreenWizardTest
Change-Id: Iede712d229289fa537b13e17241ed2174e08ff54

GitOrigin-RevId: f7919c6ed616965cf65bfa6a81b211f8232ed6f2