[Dsl] Preserve the parentDslElement in anchors
For After anchors, this is barely relevant: the parent Dsl element
should be the parent of the anchor element.  For Start anchors, this
is important: the start of what, exactly.  (Some elements, such as
GradleDslInfixExpression, will delegate anchor finding to their
parents in at least some circumstances.

This change doesn't actually use the preserved information, but does
simplify a little of the various language-specific utilities.

Bug: 158066552
Test: existing
Change-Id: I49a698abdce1cfb44ef3d19d3d1dc1d685e0c22c

GitOrigin-RevId: ea08c89a533bb910ac54c65cd3327a6a2b0ee40d