[AUA] Refactoring to block upgrade if render script property is applied
Render script property is being removed from Gradle properties in AGP
9.0. We need to block upgrades to 9.0 on projects that apply the
property (it is present and not set to no op).

This change adds a refactoring that extends
AbstractBlockPropertyUnlessNoOpProcessor to blocks upgrades that
crosses 9.0.0-alpha01 and has the property set to true or the ones that
start from 8.0.0-alpha04 or earlier. If the property is set to false,
this new refactoring will remove it from the properties file.

Bug: 375431484
Test: BlockRenderScriptPropertyPresentRefactoringProcessorTest
Change-Id: I483d9b251dc8359f28801983c6fce02084f7a73c

GitOrigin-RevId: 7cfccf6ae0169eb76893500244599d1da6da8087