Bundle performanceTesting plugin in Android Studio
It has been bundled in IntelliJ Community for a while now. We no
longer have a reason to exclude it---especially now that JetBrains
extracted the YourKit agent into a separate plugin that depends
on this one (see "YourKit Profiler for IDE Performance Testing").
If we do not bundle performanceTesting, then users have no easy way
to capture a CPU snapshot in Android Studio.

Fixes: 329416516
Change-Id: Id6098288c3e95a5f44d8180552d8bda75b1adcc8

GitOrigin-RevId: 699dccfab221110f42dafe91caecdc73bb77798a