Wrap AndroidApplicationInfo Creating with an Interface
The prod implementation depends on an external executable (AAPT) which
is not easily accessible in tests which results in a messy log warning
emitted by every ApkEditorTest test.

This allows us to use a fake in tests and reduce the noise.

It could also allow us t actually test rendering of the
AndroidApplicationInfo later on.

Bug: NA
Test: NA
Change-Id: I8a0e43ccc1a68dd72535dfc6ba48793252c1a829

GitOrigin-RevId: 34e6980b5da63e9d4ee81a8099fb4d248fde38e6