Migrate 'Install Type' step to use new wizard
The UI has been extracted from the deprecated `InstallationTypeWizardStep`
class to a new `InstallationTypeWizardStepUi` class. A new
`InstallationTypeWizardStep` class has been creatad which supports
the new `ModelWizardStep` and re-uses the existing UI.

I had to make a small change to the `SdkComponentsStep.kt` to get the
tests to pass, i.e. to check that the `InstallationTypeWizardStep`
logic was correctly updating the model. However, I'll be replacing
`SdkComponentsStep.kt` with a newly migrated class in a follow-up CL, so
I wouldn't spend much time verifying it.

You can force the new wizard to launch when running Android Studio
locally by setting the `NPW_FIRST_RUN_WIZARD` and `NPW_FIRST_RUN_SHOW`
flags to `true` in `StudioFlags.java`.

Bug: 311256502
Test: WelcomeScreenWizardTest
Change-Id: Ib91eb12e7c8851d38f6ebb6473464ca2faacfdf7

GitOrigin-RevId: 6cbd48cec44d117ab904d4e3e785410ebf38826a