Do not run queries in snapshot directories
Add `query.sync.run.query.in.workspace` experiment (default `true`) to
disable running queries in a snapshot directory.

This may result in query results reflecting a newer state of the
workspace or even failing because of inconsistencies caused by the user
editing files.

In general this is not expected to be dangerous since these changes are
supposed to be detected and the query state should become outdated
anyway and require re-querying.

Switching to a snapshot provides some consistency but at the cost of
spawning a new bazel instance and re-loading all packages every time.

Bug: 356244980
Test: n/a
Change-Id: Ib433ac2869700ff75d2c54c1bfefe1cf3ff9cc17

GitOrigin-RevId: 07b2b2286570a4f6c4d18c7cdd5b920656662c2a