Fix adding more than one dependency from the popup
Applying changes to the project model can destroy some of the internal
state: in particular, Psi elements associated with Dsl elements can
become invalid.  It's possible to reparse, but in this instance there
was no need to apply changes to the model after each added dependency;
instead, batch up the changes and apply them at the end.

Fixes: 379171292
Test: manual
Change-Id: Ibb4be7bbf872e5c63e3aa6b26f76b0d939b66982

GitOrigin-RevId: e61ac42b5a5bc38e71ffe178d1d6e2e9000257b2