Migrate 'Downloading Components' step to use new wizard
The existing `ProgressStep` class was designed to support running
long-running operations that belong to other steps. To simplify the
migration the migrated step supports running just a single
long-running task (this is all that's needed for the welcome wizard).

You can force the new wizard to launch when running Android Studio
locally by setting the `NPW_FIRST_RUN_WIZARD` and `NPW_FIRST_RUN_SHOW`
flags to `true` in `StudioFlags.java`.

Bug: 311256502
Test: WelcomeScreenWizardTest + manual
Change-Id: Ib3fd552edbf176f964954d143a44a9216b15ed07

GitOrigin-RevId: cd495c04f0f5e4d3e10024af021a36c73f9972c1