[plugin model] use a separate classloader for plugin modules with loading='required' by default (IJPL-166404)
A new option 'ModuleLoadingRule.EMBEDDED' is introduced. It has the semantics which 'REQUIRED' had before: the plugin isn't loaded if such a module isn't available, and the module is packed into the main plugin JAR and loaded by the main plugin classloader. And the semantics of 'REQUIRED' changed: the plugin isn't loaded if such a module isn't available, but the module has its own classloader and may be put to a separate JAR.

GitOrigin-RevId: 4a5b0c098eb361c15a2bbb323da78df0f7e9f99d