[FL-22733] Notify the user about missing Android SDK or required Android platforms.
The following scenarios are covered with the changes:
* No Android SDK is installed at all: we offer a fix to install SDK into the default location
* Some platforms are missing (for example, we have SDK with platform 34, but the project requires 35): we offer to download the platform in this case
* Project has an SDK in local.properties different from the one used by IDE (default or from ANDROID_HOME): in this case we rewrite local.properties and show a notification to the user
One more important thing: the way we detect Android projects was altered, also in IDEA.
Before the changes, we were looking for a top-level folder app with the AndroidManifest.xml, which is not always the case (especially with our wizard projects we typically have composeApp, not just app, folder).
After the changes, we are looking for any folder with the specific AndroidManifest.xml path (both in main and androidMain). This will allow opening more projects in IDEA initially when no local.properties file is present.

Merge-request: IJ-MR-149228
Merged-by: Ilia Bogdanovich <ilia.bogdanovich@jetbrains.com>

GitOrigin-RevId: e8cf82bed7f40f1024b41dafba705b7349a6750d
(cherry picked from commit 2dec242f247f181da7855d0b6d6713b2cd3ac523)
Change-Id: I10a44780957e2f8b991e70416d0e7b24f70dbf10