Clean up logic to show AEHD install / uninstall wizard steps
There is logic within the `Aehd` installable component to add an info
step to any wizards where it's configured to be installed. There is a
comment in `AehdInstallInfoStep` that suggests it is only added to make
sure we don't start immediately installing things without first showing
a wizard step. This makes sense in the context of `AehdWizard` but it
doesn't make sense in the context of the welcome wizard, where it adds
no value when shown.

We can simplify the logic by explicitly adding the info steps within
the `AehdWizard`. This allows us to remove the `createSteps` logic from
the `ComponentTreeNode` class and simplifies the migration to the new
wizard.

Bug: 311256502
Test: Manual
Change-Id: I160f94a87b0d22838640d7bdff86e6ed4477fe1b

GitOrigin-RevId: 998708bb7bf210ba4f3f530566b7db8cc08cd180