Fix test flakiness
The flakiness was caused by the registering of the shortcuts being done
asynchronously on the background thread.

That change was made in Ief83059055cdef04c162.

The asynchronous call was removed in I8c3a17ae71e42e5c8, however I
forgot to remove the asynchronous registering of the shortcuts. This
can now be synchronously.

Fixes: 380220695
Test: bazel test //tools/adt/idea/designer:intellij.android.designer\
 .tests_tests__designer --config=remote --runs_per_test=1000 \
 --test_filter=MultiRepresentationPreviewTest
Change-Id: I4bff4b62d690a8d450223023f3c42a7989ba961c

GitOrigin-RevId: 53e84c523c3906e1f9f67ebd4445975afa283698