[android] Ensure Gradle DSL module doesn't use classes from 'intellij.android.gradle.dsl.flags' module
Fix 'package' value of a module descriptor in the 'intellij.android.gradle.dsl.flags' module.

Gradle DSL module is a main plugin module. As a main plugin module, it shouldn't access classes of its content modules because this leads to a classloading violation, where an exception is thrown by a PluginClassloader.

GitOrigin-RevId: d2bbbe9b3731af78c35afe0df27efdd9917fe065