Prevent redundant calls to setClassGrouping for subheaps
When one selects a new item in the class grouping
(i.e. "class", "package"), a call to recompute the
heap dump table values is done via the `coalesce` method.
In the case of choosing "view all heaps", a call was done
to setClassGrouping for each of the subheaps (app, image,
zygote) in the `AllHeapSet` class, and was also done for
the parent `HeapSet` class that `AllHeapSet` inherits from.
This call to setClassGrouping in the superclass was doing a
reundant call to coalesce, and causing the values for the
heapsets to be greater than what they should. This
problematic call to the super class has been removed.

Fixes: 379429938
Test: existing
Change-Id: I685f1226742eb7fbd17c98201b4914b1cbee9954

GitOrigin-RevId: 68cca73da3d03639c21e13917bdb0cd0d2565961