Fix path to trace processor daemon binary for macOS
Change If2832797ca39c546136b8e89aeaa5c09cae64d01 split the
'darwin' folder for the trace processor daemon binary into
two different folders, one for arm and one for x86. Thus,
the current TraceProcessorDaemonManager was failing to find
the binaries for macOs becuase it only looked for the 'darwin'
directory. This change adds a check for the arch to point
to the correct, respective binary folder.

Bug: none
Test: manual
Change-Id: I68cffb6bcfbc13be6b547b6d76be1375a047d397

GitOrigin-RevId: 7ef4eeb55b599203e1a0ddd870bea384f1d53506