Implement ProjectSystemService-aware project open for ASwB
This involves admitting that the open processor is really an Android
project open processor, at least as far as the code organization is
concerned: we want to set the ID of the AndroidProjectSystem provider,
so move the open processor to `aswb` (from `base`).

Define a helper utility in ProjectSystemService in order not to have
to construct a `suspend fun` value in the open processor.

Bug: 325936093
Test: existing
Change-Id: I75a6d8b4e1e1450948d82fbdd7ae791a52353f98

GitOrigin-RevId: 5179b8bd20becd87c166feba5182d26fd5289a33