Make DFS iterative instead of recursive in AnnotationsGraph
The pre-existing recursive implementation of the DFS causes stack
overflow in some particular big cases, like the one added in
testPerformanceDeepTree in this change.

These scenarios are very unlikely to happen in reality, but it is
better to prevent it by having an iterative implementation instead
of a recursive one.

Fixes: 345737544
Test: added in AnnotationsGraphTest, and others should still pass
Change-Id: I3c4db1b7ba5dc385d47f3fc24a69b7d0fe5d0aa6

GitOrigin-RevId: 80e48693441975d745b60a115c61dea84fa5286a