Load AndroidDexer Extension
Expression evaluation works by compiling a code snippet on the fly and passing it
to the JVM to execute.

On Android, this requires Dexing the code before sending to the VM. This is done
by the AndroidDexer extension.

It seems that the test framework doesn't load plugins by itself. It may be possible
to do but it is a simple matter to register the extension manually.

This cl handles loading the extension and adds a small test target that includes
tests that depend on the extension.

Bug: NA
Test: NA
Change-Id: I3f1ea9066418b437c46074af2cb14a69acaeab45

GitOrigin-RevId: 5aaba2e81707fd49dd8118577fe36fb24ed8e6da