Run slow operation on the background
isValidComposableSet is being called on the EDT when the run
configuration is executed. The validation itself calls read actions
which can be slow. This is why we had to add the allowSlowOeprations.

This CL moves the computation to the background and wraps it in a
progress modal. The modal only shows if the computation takes a while.
In normal cases the modal doesn't show as the computation is fast.

Fixes: 322783469
Test: manually tested
Change-Id: I27d27c52382ede7fafdb8d2ea06545b53f950293

GitOrigin-RevId: 36cea920705b61d673aa3c228937e34b82a9eb84