Fix SDK components step table renderer buggy behaviour
During a previous refactor (ag/30549443) when configuring the table cell
renderer and editor, the code was updated to re-use the renderer.

This has led to buggy behaviour in the component list when selecting/de-
selecting components, i.e. sometimes you have to click twice for the
component to checkbox to update as 'selected', sometimes the cell
disappears.

Bug: 311256502
Test: Manual
Change-Id: I4987415095c34e853680396719b55ca39083262c

GitOrigin-RevId: a18020f077fbde2567bf45a81d0ae307e8977b1d