Register AndroidFieldVisibilityProvider Extension
This prevents evaluation tests from timing out when evaluating this pointer on ART.
The timeout is caused by an infinite recursion when trying to render the  shadow$_klass_ field.

A sample test case is added to exercise the code: test-shadow-fields
Note that test still fails because of a minor issue. See https://github.com/JetBrains/intellij-community/pull/2884

Bug: NA
Test: NA
Change-Id: Ia28e14dd64f108d6ca7f6505798526e96e1df283

GitOrigin-RevId: 5c156f59ee7e06475c887a6fa5369e4a9ac379bb