Remove a few deprecated usages
KotlinType.fqName is deprecated and scheduled for removal, primarily because it's K1-only and should be replaced with analysis API.

Our remaining usages of it are actually K1-only as well, as we have divergent code paths for K2. As such, I'm just inlining the property declaration since it's quite simple, and we'll be getting rid of that code in the relatively near future.

Bug: n/a
Test: n/a
Change-Id: Ib4a7f3051a30189b83645308e43cfd18f90098a9

GitOrigin-RevId: dc74b8cd9fef5da2dff346aa1c1c3d1c853d59c1